restructure dbinfo in documentation
