+++ web/restapi/__init__.py
@@ -12,6 +12,7 @@ from .search_vendor import api as search_vendor_api
 from .admin import wl as wl_api
 from .admin import bl as bl_api
 from .auth import api as auth_api
+from .dbinfo import api as dbinfo_api
 
 namespaces = [
     cpe_api,
@@ -21,6 +22,7 @@ namespaces = [
     query_api,
     browse_api,
     search_vendor_api,
+    dbinfo_api,
 ]
 
 blueprint = Blueprint("api", __name__, url_prefix="/api")
